+++ src/components/BasicInfoSteps/Photo.module.scss
@@ -6,7 +6,6 @@
   display: flex;
   flex-direction: column;
   gap: 1rem 0;
-  height: 100%;
 }
 
 .no_image {
