+++ src/components/UserCard/UserCard.module.scss
@@ -1,7 +1,8 @@
 #img {
   width: 100%;
   height: 100%;
-  background-size: contain;
+  max-height: 80vh;
+  background-size: cover;
   background-repeat: no-repeat;
   background-position: center;
 }
