+++ src/pages/Profile/AboutMe/BasicInfo/BasicInfo.module.scss
@@ -1,0 +1,24 @@
+.container {
+  display:flex;
+  flex-direction: column;
+  gap: 1rem;
+  height: 100%;
+}
+
+.content {
+  display: flex;
+  flex-direction: column;
+  gap: 2rem;
+  overflow-y: auto;
+
+  &__part{
+    display: flex;
+    flex-direction: column;
+    gap: 1rem;
+  }
+}
+
+.header {
+  display: flex;
+  justify-content: space-between;
+}
