+++ src/routes/profile/ProfileRoutes.tsx
@@ -8,9 +8,13 @@ import { settingsRoutes } from './settings/routes'
 import { questionnaireRoutes } from './questionnaire/BasicInfo'
 import QuestionnaireBasic from 'src/layouts/QuestionnaireBasic/QuestionnaireBasic'
 import { NotFound } from 'src/pages/NotFound'
+import BasicInfo from 'src/pages/Profile/AboutMe/BasicInfo/BasicInfo'
 
 export default [
-  <Route key='1' path={ProfileRoutes.ABOUT_ME} element={<AboutMe />}></Route>,
+  <Route key='1' path={ProfileRoutes.ABOUT_ME}>
+    <Route path='' element={<AboutMe />} />
+    <Route path={ProfileRoutes.BASIC_INFO} element={<BasicInfo />} />
+  </Route>,
   <Route key='2' path={ProfileRoutes.MY_APARTMENT}>
     {apartmentsRoutes}
   </Route>,
